python/requirements.txt
autogen-agentchat==0.4.7
autogen-ext[openai,azure]==0.4.7
autogen-ext[mcp]==0.4.7
fastapi
uvicorn
python-dotenv
rich
